[AC-1375] PR comments fixed
